chore: fix 48 auto-fixable warnings and 3 TypeScript errors
**TypeScript Fixes:**
- Fix broken SystemSettingsModelSelection import by inlining component
- Fix vi.fn type errors in useMessageStreaming.test.tsx
- Remove unused PlusOutlined import in ModelMappingCard.tsx

**Rust Auto-fixes (45 warnings):**
- Remove 23 unused imports across 9 crates
- Add #[derive(Default)] for 8 types (replace manual implementations)
- Simplify 7 redundant closures to function pointers
- Replace map_or with is_some_and (3 instances)
- Use strip_prefix/strip_suffix instead of manual slicing (2 instances)
- Fix redundant pattern matching and code style issues
- Fix broken test code referencing non-existent struct fields

**Results:**
- Warnings: 94 → 49 (51% reduction)
- TypeScript: Clean compilation ✓
- All tests passing ✓
- Files modified: 46

+++ crates/agent-server/src/handlers/metrics.rs
@@ -238,9 +238,9 @@ pub async fn v2_unified_summary(
 
     match (chat_result, forward_result) {
         (Ok(chat), Ok(forward)) => {
-            let total_requests = chat.total_sessions as u64 + forward.total_requests;
+            let total_requests = chat.total_sessions + forward.total_requests;
             let total_tokens = chat.total_tokens.total_tokens + forward.total_tokens.total_tokens;
-            let total_success = (chat.total_sessions - chat.active_sessions) as u64
+            let total_success = (chat.total_sessions - chat.active_sessions)
                 + forward.successful_requests;
             let total_errors = forward.failed_requests;
             let success_rate = if total_requests > 0 {
